Project Update
This is a clone of the project source as it is on our (MMOD) internal
servers with a bit of stripping having been performed.

- Extends the MVC framework by providing both 'front' and 'back' end
applications invocable from the same directory and which work as
separate applications for tighter security.

- Demonstrates proper usage of nk and nk-mysql as of their most recent
versions.

- A skeleton for an extension system is present for both the 'front' and
'back' end applications though it is non-functional.

- Examples of both a 'dashboard' and 'website' are available by visiting
the home page and /dash at the application URL when running.

- Install directory contains an sql file that will need modification in
order to set up a user authentication system and support for basic
facilities. Advanced cryptography methods are removed from this variant
so it is advised to implement some added security of your own when
designing how you will encrypt, store, and fetch data appropriately for
your own authentication system. Please see /app/models/account and
/adm/models/account respectively.

- A full AJAX Client Framework is provided

- At this stage not everything is fully functional, please bear with me
as I get things updated and sorted and prepare documentation for
installation and deployment.